% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{components}
\alias{components}
\alias{necessary_causes}
\alias{sufficient_causes}
\title{Identify components of a set of sufficient causes}
\usage{
components(causes)

necessary_causes(causes)

sufficient_causes(causes)
}
\arguments{
\item{causes}{a data frame, the result of \code{\link[=causify]{causify()}}}
}
\value{
a character vector
}
\description{
\code{components()} identifies the unique components across all sufficient causes.
\code{necessary_causes()} identifies the components that appear in all sufficient
causes. \code{sufficient_causes()} identifies all the components of a sufficient
cause.
}
\examples{

causes <- causify(sc(A = 1, B = 0), sc(A = 1, E = 1, C = 0))

components(causes)
necessary_causes(causes)
sufficient_causes(causes)

}
