% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalpie.R
\name{pies}
\alias{pies}
\alias{causal_pie}
\alias{causal_pie_necessary}
\title{Causal Pies}
\usage{
causal_pie(causes, text_col = "black")

causal_pie_necessary(causes, text_col = "black")
}
\arguments{
\item{causes}{a data frame, the result of \code{\link[=causify]{causify()}}}

\item{text_col}{color of the text. Default is "black".}
}
\value{
a ggplot
}
\description{
Causal pies are visual representations of sufficient causes. \code{causal_pie()}
plots all sufficient causes and highlights the unique components.
\code{causal_pie_necessary()} does the same but highlights necessary causes (those
components that appear in every sufficient cause).
}
\examples{
causes <- causify(sc(A = 1, B = 0), sc(A = 1, E = 1, C = 0))

causes |>
  causal_pie() +
  theme_causal_pie()

causes |>
  causal_pie_necessary() +
  theme_causal_pie_grid()

}
